+++ packages/af-utils-react-virtual-headless/src/components/Subscription/index.js
@@ -1,14 +1,21 @@
-import { memo, useCallback } from "react";
+import { useCallback } from "react";
 import { useSyncExternalStore } from "use-sync-external-store/shim";
 import { EVT_ALL } from "constants";
 
-const Subscription = ({ model, children, getHash, events = EVT_ALL }) => (
+/* TODO: dirty; based on evt constants. Maybe macro to evaluate build-time? */
+const EVT_TO_PROP = ["from", "to", "scrollSize"];
+
+const Subscription = ({ model, children, events = EVT_ALL }) => {
+    const getHash = () =>
+        events.reduce((acc, e) => `${acc}_${model[EVT_TO_PROP[e]]}`, "");
+
     useSyncExternalStore(
-        useCallback(listener => model.on(listener, events)[events]),
+        useCallback(listener => model.on(listener, events), [events]),
         getHash,
         getHash
-    ),
-    children(model)
-);
+    );
 
-export default memo(Subscription);
+    return children();
+};
+
+export default Subscription;