+++ packages/af-utils-react-virtual-headless/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @af-utils/react-virtual-headless
 
+## 0.0.14
+
+### Patch Changes
+
+-   removed getHash, calculating it automatically now
+
 ## 0.0.13
 
 ### Patch Changes